FEAT: auth logic for taskservice

+++ api/src/main/java/io/mattinfern0/kanbanboardapi/tasks/TaskService.java
@@ -14,11 +14,15 @@ import io.mattinfern0.kanbanboardapi.tasks.dtos.CreateUpdateTaskDto;
 import io.mattinfern0.kanbanboardapi.tasks.dtos.TaskDetailDto;
 import io.mattinfern0.kanbanboardapi.tasks.dtos.UpdateTaskColumnPositionDTO;
 import io.mattinfern0.kanbanboardapi.tasks.mappers.TaskDtoMapper;
+import io.mattinfern0.kanbanboardapi.users.UserAccessService;
+import jakarta.annotation.Nullable;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.access.AccessDeniedException;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.security.Principal;
 import java.util.List;
 import java.util.UUID;
 
@@ -32,8 +36,10 @@ public class TaskService {
     final OrganizationRepository organizationRepository;
     final UserRepository userRepository;
     final TaskStatusService taskStatusService;
+    final UserAccessService userAccessService;
 
     final TaskDtoMapper taskDtoMapper;
+
 
     @Autowired
     public TaskService(
@@ -42,36 +48,70 @@ public class TaskService {
         OrganizationRepository organizationRepository,
         UserRepository userRepository,
         TaskStatusService taskStatusService,
-        TaskDtoMapper taskDtoMapper
-    ) {
+        TaskDtoMapper taskDtoMapper,
+        UserAccessService userAccessService) {
         this.taskRepository = taskRepository;
         this.boardColumnRepository = boardColumnRepository;
         this.organizationRepository = organizationRepository;
         this.userRepository = userRepository;
         this.taskStatusService = taskStatusService;
         this.taskDtoMapper = taskDtoMapper;
+        this.userAccessService = userAccessService;
     }
 
-    public List<TaskDetailDto> getTaskList() {
-        List<Task> taskEntities = taskRepository.findAll();
+    public List<TaskDetailDto> getTaskList(Principal principal, @Nullable UUID organizationId) {
+        if (organizationId == null) {
+            return List.of();
+        }
+
+        if (!userAccessService.canAccessOrganization(principal, organizationId)) {
+            throw new AccessDeniedException("You do not have permission to access this resource");
+        }
+
+        List<Task> taskEntities = taskRepository.findByOrganizationId(organizationId);
         return taskDtoMapper.taskListToTaskDetailDtoList(taskEntities);
     }
 
-    public TaskDetailDto getTaskDetail(UUID taskId) {
+    public TaskDetailDto getTaskDetail(Principal principal, UUID taskId) {
         Task entity = taskRepository.findById(taskId)
             .orElseThrow(() -> new ResourceNotFoundException(String.format("Task with id %s not found", taskId)));
         return taskDtoMapper.taskToTaskDetailDto(entity);
     }
 
     @Transactional
-    public TaskDetailDto createTask(@Valid CreateUpdateTaskDto createUpdateTaskDto) {
+    public TaskDetailDto createTask(Principal principal, @Valid CreateUpdateTaskDto createUpdateTaskDto) {
+        authorizeCreateUpdateTask(principal, createUpdateTaskDto);
+
         Task newTask = taskFromCreateTaskDto(createUpdateTaskDto);
         taskRepository.save(newTask);
         return taskDtoMapper.taskToTaskDetailDto(newTask);
     }
 
+    void authorizeCreateUpdateTask(Principal principal, CreateUpdateTaskDto createUpdateTaskDto) {
+        if (
+            createUpdateTaskDto.organizationId() != null
+                && !userAccessService.canAccessOrganization(principal, createUpdateTaskDto.organizationId())
+        ) {
+            throw new AccessDeniedException("You don't have permission to create tasks in this organization");
+        }
+
+        if (
+            createUpdateTaskDto.boardColumnId() != null
+        ) {
+            BoardColumn column = boardColumnRepository
+                .findById(createUpdateTaskDto.boardColumnId())
+                .orElseThrow(() -> new ResourceNotFoundException(
+                    String.format("BoardColumn with id %s not found", createUpdateTaskDto.boardColumnId())
+                ));
+
+            if (!userAccessService.canAccessBoard(principal, column.getBoard().getId())) {
+                throw new AccessDeniedException("You don't have permission to create tasks in this board");
+            }
+        }
+    }
+
     @Transactional
-    public TaskDetailDto updateTask(UUID taskId, @Valid CreateUpdateTaskDto dto) {
+    public TaskDetailDto updateTask(Principal principal, UUID taskId, @Valid CreateUpdateTaskDto dto) {
         Task task = taskRepository
             .findById(taskId)
             .orElseThrow(() -> new ResourceNotFoundException(String.format("Task with id %s not found", taskId)));
@@ -81,6 +121,12 @@ public class TaskService {
             .orElseThrow(() -> new ResourceNotFoundException(
                 String.format("Organization with id %s not found", dto.organizationId())
             ));
+
+        if (!userAccessService.canAccessTask(principal, taskId)) {
+            throw new AccessDeniedException("You do not have permission to update this task");
+        }
+
+        authorizeCreateUpdateTask(principal, dto);
 
         task.setDescription(dto.description());
         task.setTitle(dto.title());
@@ -119,10 +165,14 @@ public class TaskService {
     }
 
     @Transactional
-    public void updateTaskAssignees(UUID taskId, List<UUID> assigneeIds) {
+    public void updateTaskAssignees(Principal principal, UUID taskId, List<UUID> assigneeIds) {
         Task task = taskRepository
             .findById(taskId)
             .orElseThrow(() -> new ResourceNotFoundException(String.format("Task with id %s not found", taskId)));
+
+        if (!userAccessService.canAccessTask(principal, taskId)) {
+            throw new AccessDeniedException("You do not have permission to update this task");
+        }
 
         if (assigneeIds.isEmpty()) {
             task.getAssignees().clear();
@@ -136,16 +186,21 @@ public class TaskService {
         taskRepository.saveAndFlush(task);
     }
 
-    public void deleteTask(UUID taskId) {
+    public void deleteTask(Principal principal, UUID taskId) {
         if (!taskRepository.existsById(taskId)) {
             throw new ResourceNotFoundException(String.format("Task with id %s not found", taskId));
         }
+
+        if (!userAccessService.canAccessTask(principal, taskId)) {
+            throw new AccessDeniedException("You do not have permission to delete this task");
+        }
+
         taskRepository.deleteById(taskId);
     }
 
 
     @Transactional
-    public void updateTaskColumnPosition(UUID taskId, UpdateTaskColumnPositionDTO dto) {
+    public void updateTaskColumnPosition(Principal principal, UUID taskId, UpdateTaskColumnPositionDTO dto) {
         Task task = taskRepository
             .findById(taskId)
             .orElseThrow(() -> new ResourceNotFoundException(String.format("Task with id %s not found", taskId)));
